Move retrieval QA serializers to retrieval_qa.py
